+++ crux_core/src/testing.rs
@@ -90,6 +90,21 @@ where
         Ok(self.context.updates(Vec::new()))
     }
 
+    /// Resolve an effect `request` from previous update, then run the resulting event
+    ///
+    /// This helper is useful for the common case where  one expects the effect to resolve
+    /// to exactly one event, which should then be run by the app.
+    pub fn resolve_to_event_then_update<Op: Operation>(
+        &self,
+        request: &mut Request<Op>,
+        value: Op::Output,
+        model: &mut App::Model,
+    ) -> Update<Ef, App::Event> {
+        request.resolve(value).expect("failed to resolve request");
+        let event = self.context.updates().expect_one_event();
+        self.update(event, model)
+    }
+
     /// Run the app's `view` function with a model state
     pub fn view(&self, model: &App::Model) -> App::ViewModel {
         self.app.view(model)
@@ -146,6 +161,7 @@ impl<Ef, Ev> AppContext<Ef, Ev> {
 /// Update test helper holds the result of running an app update using [`AppTester::update`]
 /// or resolving a request with [`AppTester::resolve`].
 #[derive(Debug)]
+#[must_use]
 pub struct Update<Ef, Ev> {
     /// Effects requested from the update run
     pub effects: Vec<Ef>,
@@ -164,6 +180,70 @@ impl<Ef, Ev> Update<Ef, Ev> {
 
     pub fn effects_mut(&mut self) -> impl Iterator<Item = &mut Ef> {
         self.effects.iter_mut()
+    }
+
+    /// Assert that the update contains exactly one effect and zero events,
+    /// and return the effect
+    pub fn expect_one_effect(mut self) -> Ef {
+        if self.events.is_empty() && self.effects.len() == 1 {
+            self.effects.pop().unwrap()
+        } else {
+            panic!(
+                "Expected one effect but found {} effect(s) and {} event(s)",
+                self.effects.len(),
+                self.events.len()
+            );
+        }
+    }
+
+    /// Assert that the update contains exactly one event and zero effects,
+    /// and return the event
+    pub fn expect_one_event(mut self) -> Ev {
+        if self.effects.is_empty() && self.events.len() == 1 {
+            self.events.pop().unwrap()
+        } else {
+            panic!(
+                "Expected one event but found {} effect(s) and {} event(s)",
+                self.effects.len(),
+                self.events.len()
+            );
+        }
+    }
+
+    /// Assert that the update contains no effects or events
+    pub fn assert_empty(self) {
+        if self.effects.is_empty() && self.events.is_empty() {
+            return;
+        }
+        panic!(
+            "Expected empty update but found {} effect(s) and {} event(s)",
+            self.effects.len(),
+            self.events.len()
+        );
+    }
+
+    /// Take effects matching the `predicate` out of the [`Update`]
+    /// and return them, mutating the `Update`
+    pub fn take_effects<P>(&mut self, predicate: P) -> VecDeque<Ef>
+    where
+        P: FnMut(&Ef) -> bool,
+    {
+        let (matching_effects, other_effects) = self.take_effects_partitioned_by(predicate);
+
+        self.effects = other_effects.into_iter().collect();
+
+        matching_effects
+    }
+
+    /// Take all of the effects out of the [`Update`]
+    /// and split them into those matching `predicate` and the rest
+    pub fn take_effects_partitioned_by<P>(&mut self, predicate: P) -> (VecDeque<Ef>, VecDeque<Ef>)
+    where
+        P: FnMut(&Ef) -> bool,
+    {
+        std::mem::take(&mut self.effects)
+            .into_iter()
+            .partition(predicate)
     }
 }