+++ crux_platform/src/lib.rs
@@ -5,7 +5,7 @@ use crux_core::macros::Capability;
 use crux_core::Command;
 use serde::{Deserialize, Serialize};
 
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
+#[derive(Clone, Debug, Default, PartialEq, Eq, Serialize, Deserialize)]
 pub struct PlatformRequest;
 
 // TODO revisit this