+++ examples/cat_facts/shared/src/app.rs
@@ -157,16 +157,7 @@ impl App for CatFacts {
             }
             Event::SetFact(Ok(mut response)) => {
                 model.cat_fact = Some(response.take_body().unwrap());
-
-                let bytes = serde_json::to_vec(&model).unwrap();
-                Command::effect({
-                    let key_value = caps.key_value.clone();
-                    async move {
-                        let _ = key_value.set_async(KEY.to_string(), bytes).await;
-                        Command::none()
-                    }
-                })
-                .join(caps.time.now(Event::CurrentTime))
+                caps.time.now(Event::CurrentTime)
             }
             Event::SetImage(Ok(mut response)) => {
                 model.cat_image = Some(response.take_body().unwrap());
@@ -189,6 +180,7 @@ impl App for CatFacts {
             Event::CurrentTime(TimeResponse::Now(instant)) => {
                 let time: DateTime<Utc> = instant.try_into().unwrap();
                 model.time = Some(time.to_rfc3339_opts(chrono::SecondsFormat::Secs, true));
+
                 let bytes = serde_json::to_vec(&model).unwrap();
 
                 Command::effect({
@@ -242,83 +234,109 @@ impl App for CatFacts {
 
 #[cfg(test)]
 mod tests {
-    use assert_let_bind::assert_let;
     use crux_core::testing::AppTester;
     use crux_http::{
         protocol::{HttpRequest, HttpResponse, HttpResult},
         testing::ResponseBuilder,
     };
-
-    use crate::Effect;
+    use crux_kv::{value::Value, KeyValueOperation, KeyValueResponse, KeyValueResult};
+    use crux_time::Instant;
 
     use super::*;
-
-    #[test]
-    fn fetch_sends_some_requests() {
-        let app = AppTester::<CatFacts, _>::default();
-        let mut model = Model::default();
-
-        let update = app.update(Event::Fetch, &mut model);
-
-        assert_let!(Effect::Http(request), update.effects().next().unwrap());
-        let actual = &request.operation;
-        let expected = &HttpRequest::get(FACT_API_URL).build();
-
-        assert_eq!(actual, expected);
-    }
 
     #[test]
     fn fetch_results_in_set_fact_and_set_image() {
         let app = AppTester::<CatFacts, _>::default();
         let mut model = Model::default();
 
-        let mut update = app.update(Event::Fetch, &mut model);
-        let mut effects = update.effects_mut();
+        // send fetch event to app
+        let (mut http_effects, mut render_effects) = app
+            .update(Event::Fetch, &mut model)
+            .take_effects_partitioned_by(Effect::is_http);
 
-        assert_let!(Effect::Http(request), effects.next().unwrap());
-        let actual = &request.operation;
-        let expected = &HttpRequest::get(FACT_API_URL).build();
-        assert_eq!(actual, expected);
+        // receive render effect
+        render_effects.pop_front().unwrap().expect_render();
+
+        // receive two HTTP effects, one to fetch the fact and one to fetch the image
+        // we'll handle the fact request first
+        let request = &mut http_effects.pop_front().unwrap().expect_http();
+        assert_eq!(request.operation, HttpRequest::get(FACT_API_URL).build());
 
         let a_fact = CatFact {
             fact: "cats are good".to_string(),
             length: 13,
         };
 
-        let response = HttpResult::Ok(
-            HttpResponse::ok()
-                .body(r#"{ "fact": "cats are good", "length": 13 }"#)
-                .build(),
+        // resolve the request with a simulated response from the web API
+        let event = app
+            .resolve(
+                request,
+                HttpResult::Ok(
+                    HttpResponse::ok()
+                        .body(r#"{ "fact": "cats are good", "length": 13 }"#)
+                        .build(),
+                ),
+            )
+            .expect("should resolve successfully")
+            .expect_one_event();
+
+        // check that the app emitted an (internal) event to update the model
+        assert_eq!(
+            event,
+            Event::SetFact(Ok(ResponseBuilder::ok().body(a_fact.clone()).build()))
         );
-        let update = app
+
+        // Setting the fact should trigger a time event
+        let mut time_events = app.update(event, &mut model).take_effects(|e| e.is_time());
+
+        let request = &mut time_events.pop_front().unwrap().expect_time();
+
+        let response = TimeResponse::Now(Instant::new(0, 0).unwrap());
+        let event = app
             .resolve(request, response)
-            .expect("should resolve successfully");
+            .expect("should resolve successfully")
+            .expect_one_event();
 
-        let expected_response = ResponseBuilder::ok().body(a_fact.clone()).build();
-        assert_eq!(update.events, vec![Event::SetFact(Ok(expected_response))]);
+        assert_eq!(event, Event::CurrentTime(response));
 
-        for event in update.events {
-            app.update(event, &mut model);
-        }
+        // update the app with the current time event
+        // and check that we get a key value set event and a render event
+        let (mut key_value_effects, mut render_effects) = app
+            .update(event, &mut model)
+            .take_effects_partitioned_by(Effect::is_key_value);
+        render_effects.pop_front().unwrap().expect_render();
 
-        assert_let!(Effect::Http(request), effects.next().unwrap());
-        let actual = &request.operation;
-        let expected = &HttpRequest::get(IMAGE_API_URL).build();
-        assert_eq!(actual, expected);
+        let request = &mut key_value_effects.pop_front().unwrap().expect_key_value();
+        assert_eq!(
+            request.operation,
+            KeyValueOperation::Set {
+                key: KEY.to_string(),
+                value: serde_json::to_vec(&model).unwrap()
+            }
+        );
 
-        let a_image = CatImage {
+        let _updated = app.resolve_to_event_then_update(
+            request,
+            KeyValueResult::Ok {
+                response: KeyValueResponse::Set {
+                    previous: Value::None,
+                },
+            },
+            &mut model,
+        );
+
+        // Now we'll handle the image
+        let request = &mut http_effects.pop_front().unwrap().expect_http();
+        assert_eq!(request.operation, HttpRequest::get(IMAGE_API_URL).build());
+
+        let an_image = CatImage {
             href: "image_url".to_string(),
         };
 
         let response = HttpResult::Ok(HttpResponse::ok().body(r#"{"href":"image_url"}"#).build());
-        let update = app
-            .resolve(request, response)
-            .expect("should resolve successfully");
-        for event in update.events {
-            app.update(event, &mut model);
-        }
+        let _updated = app.resolve_to_event_then_update(request, response, &mut model);
 
         assert_eq!(model.cat_fact, Some(a_fact));
-        assert_eq!(model.cat_image, Some(a_image));
+        assert_eq!(model.cat_image, Some(an_image));
     }
 }