+++ crux_core/CHANGELOG.md
@@ -8,6 +8,21 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [0.10.0](https://github.com/redbadger/crux/compare/crux_core-v0.9.1...crux_core-v0.10.0) - 2024-10-23
+
+Several additional methods to help with testing Crux apps:
+
+- Adds a `Clone` bound on  the `Operation` trait so that we can examine the operation and still resolve
+  its owning request later on — this is a breaking change.
+- Adds a `take_effects` method on `Update` to allow you to take effects off the Update that match the predicate
+- Adds a `take_effects_partitioned_by` method on `Update` to allow you to take effects off the Update
+  that match the predicate and also the remaining effects that don't match
+
+## [0.9.1](https://github.com/redbadger/crux/compare/crux_core-v0.9.0...crux_core-v0.9.1) - 2024-10-21
+
+- Fixes a memory leak that affects tasks that contain futures, see https://github.com/redbadger/crux/issues/268.
+  This only affects tasks that contain futures, and only when the task is dropped before the future completes.
+
 ## [0.9.0](https://github.com/redbadger/crux/compare/crux_core-v0.8.1...crux_core-v0.9.0) - 2024-08-13
 
 ### Other